requirements.txt
streamlit
google-genai
langchain-google-genai
langchain
pypdf
faiss-cpu